cdk/requirements.txt
aws-cdk-lib==2.3.0
